Fix stale model test fixtures
